Refine the code structure of dygraph_to_static test=develop
